aider: Make the displayed word more prominent and display an image related to the word when the match is successful.
# Aider chat conversation:

USER: make the displayed word more prominant. also when the match is successful display an image relating to the word

ASSISTANT: To make the displayed word more prominent, we need to modify `styles.css` to increase the font size and weight of the `#word` element.

To display an image related to the word when the match is successful, we need to modify `index.html` to add an `img` element with an id of "word-image", and `form.js` to update the `src` attribute of the `img` element when the match is successful.

Please note that for the image to display correctly, the entered string must correspond to the name of an image file in the same directory as the HTML file. If the image file is in a different directory or has a different name, the image will not display.

Here are the *SEARCH/REPLACE* blocks:

styles.css
```css
<<<<<<< SEARCH
#success-message {
    color: green;
    font-weight: bold;
}
=======
#word {
    font-size: 2em;
    font-weight: bold;
}

#success-message {
    color: green;
    font-weight: bold;
}
>>>>>>> REPLACE
```

index.html
```html
<<<<<<< SEARCH
    <p id="word"></p>
    <p id="success-message" style="color: green;"></p>
=======
    <p id="word"></p>
    <img id="word-image" style="display: none;" />
    <p id="success-message" style="color: green;"></p>
>>>>>>> REPLACE
```

form.js
```javascript
<<<<<<< SEARCH
                if (sortedWord === string) {
                    document.getElementById('success-message').textContent = 'Success!';
                }
=======
                if (sortedWord === string) {
                    document.getElementById('success-message').textContent = 'Success!';
                    var wordImage = document.getElementById('word-image');
                    wordImage.src = string + '.jpg';
                    wordImage.style.display = 'block';
                }
>>>>>>> REPLACE
```