aider: The word image is now based on the entered string using the Unsplash API.
# Aider chat conversation:

USER: the images should be based on the word. they are currently random

ASSISTANT: To make this change, we need to modify `form.js`. We will replace the Lorem Picsum URL with a URL from the Unsplash API that includes the entered string as a search query. This will return a random image related to the entered string.

Here is the *SEARCH/REPLACE* block:

form.js
```javascript
<<<<<<< SEARCH
wordImage.src = 'https://picsum.photos/200';
=======
wordImage.src = `https://source.unsplash.com/200x200/?${string}`;
>>>>>>> REPLACE
```

Please note that the Unsplash API does not require an access key for this type of request, but it is subject to rate limiting. If you make too many requests in a short period of time, you may be temporarily blocked from making further requests.